requirements.txt
Flask>=3.0
PyMySQL>=1.1
python-dotenv>=1.0
DBUtils>=3.1
Flask-Mail>=0.10
Werkzeug>=3.0
Flask-WTF>=1.2
Flask-Limiter>=3.5
